+++ src/Tickets/Commerce/Order.php
@@ -1237,6 +1237,8 @@ class Order extends Abstract_Order {
 		 */
 		do_action( 'tec_tickets_commerce_order_checkout_completed', $order_id );
 
+		as_unschedule_action( 'tec_tickets_commerce_async_webhook_process', [ 'order_id' => $order_id ], 'tec-tickets-commerce-stripe-webhooks' );
+
 		return (bool) as_enqueue_async_action(
 			'tec_tickets_commerce_async_webhook_process',
 			[
@@ -1312,6 +1314,15 @@ class Order extends Abstract_Order {
 		 */
 		do_action( 'tec_tickets_commerce_order_on_checkout_screen_hold_set', $order_id, $on_screen_hold );
 
+		as_unschedule_action(
+			'tec_tickets_commerce_async_webhook_process',
+			[
+				'order_id' => $order_id,
+				'try'      => 0,
+			],
+			'tec-tickets-commerce-stripe-webhooks'
+		);
+
 		return (bool) as_schedule_single_action(
 			$on_screen_hold + MINUTE_IN_SECONDS, // We schedule the action to run after the timeout.
 			'tec_tickets_commerce_async_webhook_process',
@@ -1346,6 +1357,15 @@ class Order extends Abstract_Order {
 		 * @param int $order_id The order ID.
 		 */
 		do_action( 'tec_tickets_commerce_order_on_checkout_screen_hold_remove', $order_id );
+
+		as_unschedule_action(
+			'tec_tickets_commerce_async_webhook_process',
+			[
+				'order_id' => $order_id,
+				'try'      => 0,
+			],
+			'tec-tickets-commerce-stripe-webhooks'
+		);
 
 		return (bool) as_schedule_single_action(
 			time(),