+++ src/Tribe/Attendees_Table.php
@@ -947,6 +947,15 @@ class Tribe__Tickets__Attendees_Table extends WP_List_Table {
 			return;
 		}
 
+		/** @var Tribe__Tickets__Attendees $attendees */
+		$attendees = tribe( 'tickets.attendees' );
+
+		$post = get_post( $this->get_post_id() );
+
+		if ( ! $attendees->user_can_manage_attendees( 0, $post->ID ?? '' ) ) {
+			return;
+		}
+
 		foreach ( $attendee_ids as $attendee ) {
 			list( $id, $addon ) = $this->attendee_reference( $attendee );
 
@@ -958,7 +967,6 @@ class Tribe__Tickets__Attendees_Table extends WP_List_Table {
 		}
 
 		// Redirect after deleting attendees back to attendee URL.
-		$post = get_post( $this->get_post_id() );
 		if ( ! isset( $post->ID ) ) {
 			return;
 		}
@@ -967,7 +975,7 @@ class Tribe__Tickets__Attendees_Table extends WP_List_Table {
 			return;
 		}
 
-		$redirect_url = tribe( 'tickets.attendees' )->get_report_link( $post );
+		$redirect_url = $attendees->get_report_link( $post );
 		wp_safe_redirect( $redirect_url );
 
 		exit;