+++ src/views/v2/commerce/checkout/purchaser-info.php
@@ -10,8 +10,9 @@
  * @link    https://evnt.is/1amp Help article for RSVP & Ticket template files.
  *
  * @since 5.3.0
+ * @since 5.19.3 Added a surrounding form to the purchaser info fields.
  *
- * @version 5.3.0
+ * @version 5.19.3
  *
  * @var Checkout_Shortcode $shortcode [Global] The checkout shortcode instance.
  * @var \Tribe__Template $this [Global] Template object.
@@ -29,26 +30,28 @@ if ( ! $shortcode->should_display_purchaser_info() ) {
 ?>
 <div class="tribe-tickets__form tribe-tickets__commerce-checkout-purchaser-info-wrapper tribe-common-b2">
 	<h4 class="tribe-common-h5 tribe-tickets__commerce-checkout-purchaser-info-title"><?php echo esc_html( $shortcode->get_purchaser_info_title() ); ?></h4>
-	<?php
-	$this->template(
-		'checkout/purchaser-info/name',
-		[
-			'show_address' => $shortcode->should_display_billing_info(),
-			'field'        => $billing_fields['name'],
-		]
-	);
-	?>
-	<?php $this->template( 'checkout/purchaser-info/email', [ 'field' => $billing_fields['email'] ] ); ?>
-	<?php if ( $shortcode->should_display_billing_info() ) : ?>
-		<?php $this->template( 'checkout/purchaser-info/address', [ 'field' => $billing_fields['address'] ] ); ?>
-		<div class="tribe-tickets__commerce-checkout-address-wrapper">
-			<?php $this->template( 'checkout/purchaser-info/city', [ 'field' => $billing_fields['city'] ] ); ?>
-			<?php $this->template( 'checkout/purchaser-info/state', [ 'field' => $billing_fields['state'] ] ); ?>
-			<?php $this->template( 'checkout/purchaser-info/zip', [ 'field' => $billing_fields['zip'] ] ); ?>
-			<?php $this->template( 'checkout/purchaser-info/country', [ 'field' => $billing_fields['country'] ] ); ?>
-		</div>
-		<button id="tec-tc-gateway-stripe-render-payment" class="tribe-common-c-btn tribe-tickets__commerce-checkout-form-submit-button">
-			<?php esc_html_e( 'Proceed to payment', 'event-tickets' ); ?>
-		</button>
-	<?php endif; ?>
+	<form class="tribe-tickets__commerce-checkout-purchaser-info-wrapper__form">
+		<?php
+		$this->template(
+			'checkout/purchaser-info/name',
+			[
+				'show_address' => $shortcode->should_display_billing_info(),
+				'field'        => $billing_fields['name'],
+			]
+		);
+		?>
+		<?php $this->template( 'checkout/purchaser-info/email', [ 'field' => $billing_fields['email'] ] ); ?>
+		<?php if ( $shortcode->should_display_billing_info() ) : ?>
+			<?php $this->template( 'checkout/purchaser-info/address', [ 'field' => $billing_fields['address'] ] ); ?>
+			<div class="tribe-tickets__commerce-checkout-address-wrapper">
+				<?php $this->template( 'checkout/purchaser-info/city', [ 'field' => $billing_fields['city'] ] ); ?>
+				<?php $this->template( 'checkout/purchaser-info/state', [ 'field' => $billing_fields['state'] ] ); ?>
+				<?php $this->template( 'checkout/purchaser-info/zip', [ 'field' => $billing_fields['zip'] ] ); ?>
+				<?php $this->template( 'checkout/purchaser-info/country', [ 'field' => $billing_fields['country'] ] ); ?>
+			</div>
+			<button id="tec-tc-gateway-stripe-render-payment" class="tribe-common-c-btn tribe-tickets__commerce-checkout-form-submit-button">
+				<?php esc_html_e( 'Proceed to payment', 'event-tickets' ); ?>
+			</button>
+		<?php endif; ?>
+	</form>
 </div>