+++ src/views/v2/tickets.php
@@ -10,8 +10,9 @@
  * @link    https://evnt.is/1amp Help article for RSVP & Ticket template files.
  *
  * @since   5.0.3
+ * @since 5.20.0 Added waitlist entry point.
  *
- * @version 5.0.3
+ * @version 5.20.0
  *
  * @var Tribe__Tickets__Editor__Template   $this                        [Global] Template object.
  * @var Tribe__Tickets__Tickets            $provider                    [Global] The tickets provider class.
@@ -79,6 +80,7 @@ $classes = [
 
 		<?php $this->template( 'v2/components/loader/loader' ); ?>
 
+		<?php $this->do_entry_point( 'etp-waitlist' ); ?>
 	</form>
 
 	<?php