+++ tests/_support/Commerce/Ticket_Maker.php
@@ -2,8 +2,11 @@
 
 namespace Tribe\Tickets\Test\Commerce;
 
+use Exception;
 use Tribe__Tickets__Global_Stock as Global_Stock;
 use Tribe__Tickets__Tickets_Handler;
+use Tribe__Tickets__Tickets as Tickets;
+use Tribe__Tickets__Ticket_Object as Ticket_Object;
 
 trait Ticket_Maker {
 
@@ -27,7 +30,6 @@ trait Ticket_Maker {
 			'ticket_name'             => "Test ticket for {$post_id}",
 			'ticket_description'      => "Test ticket description for {$post_id}",
 			'ticket_show_description' => 1,
-			'ticket_price'            => absint( $price ),
 			'ticket_start_date'       => '2020-01-02',
 			'ticket_start_time'       => '08:00:00',
 			'ticket_end_date'         => '2050-03-01',
@@ -39,13 +41,41 @@ trait Ticket_Maker {
 			],
 		];
 
+		if ( null !== $price ) {
+			$data['ticket_price'] = absint( $price );
+		}
+
 		$data = array_merge( $data, $overrides );
 
 		/** @var Tribe__Tickets__Tickets_Handler $tickets_handler */
 		$tickets_handler = tribe( 'tickets.handler' );
 		update_post_meta( $post_id, $tickets_handler->key_provider_field, get_class( $provider_class ) );
 
-		return $provider_class->ticket_add( $post_id, $data );
+		$id = $provider_class->ticket_add( $post_id, $data );
+
+		wp_cache_flush_group( 'post-queries' );
+		return $id;
+	}
+
+	/**
+	 * Update a ticket.
+	 *
+	 * @param int   $ticket_id The ID of the ticket to update.
+	 * @param array $overrides An array of values to override the default and random generation arguments.
+	 *
+	 * @return int|false The new ticket ID or false if not saved.
+	 *
+	 * @throws Exception If the ticket is not found.
+	 */
+	protected function update_ticket( $ticket_id, array $overrides ) {
+		wp_cache_flush_group( 'post-queries' );
+		$ticket = Tickets::load_ticket_object( $ticket_id );
+		if ( ! $ticket instanceof Ticket_Object ) {
+			throw new Exception( 'Ticket not found!' );
+		}
+
+		$overrides = array_merge( $overrides, [ 'ticket_id' => $ticket->ID ] );
+		return $this->create_ticket( $ticket->provider_class, $ticket->get_event_id(), null, $overrides );
 	}
 
 	/**